gui/requirements.txt
pycryptodome==3.9.8
checksumdir==1.2.0